backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.0
